refactor: split 7 more large files by concern (HV round 2)
webhook/webhook.go (1070 → 4 files):
- webhook.go: types, CRUD
- webhook_delivery.go: delivery and retry logic
- webhook_trigger.go: trigger management
- webhook_crypto.go: signatures, validation, private IP checks

api/tenant_handler.go (1068 → 3 files):
- tenant_handler.go: tenant CRUD, helpers
- tenant_handler_admin.go: admin assignment methods
- tenant_handler_schema.go: declarative schema methods

realtime/subscription.go (1030 → 3 files):
- subscription.go: SubscriptionManager core
- subscription_rls.go: RLS cache
- subscription_db.go: SubscriptionDB interface and impl

ai/user_kb_handler.go (1118 → 4 files):
- user_kb_handler.go: KB CRUD, permissions
- user_kb_handler_documents.go: document methods
- user_kb_handler_search.go: search methods
- user_kb_handler_entities.go: entity/graph methods

api/ddl_handler.go (1014 → 3 files):
- ddl_handler.go: core, validation, schema ops
- ddl_handler_table.go: table DDL methods
- ddl_handler_column.go: column DDL methods

mcp/tools/ddl.go (1087 → 3 files):
- ddl.go: shared helpers, schema tools
- ddl_table.go: table tools
- ddl_column.go: column tools

mcp/tools/branching.go (1034 → 3 files):
- branching_lifecycle.go: create/delete/reset/set-active
- branching_access.go: list/get/grant/revoke
