+++ internal/webhook/webhook.go
@@ -1,20 +1,11 @@
 package webhook
 
 import (
-	"bytes"
 	"context"
-	"crypto/hmac"
-	"crypto/sha256"
-	"encoding/hex"
 	"encoding/json"
 	"errors"
 	"fmt"
-	"io"
-	"net"
 	"net/http"
-	"net/url"
-	"strconv"
-	"strings"
 	"sync"
 	"time"
 
@@ -96,150 +87,6 @@ func (s *WebhookService) AllowPrivateIPs() bool {
 	return s.allowPrivateIPs
 }
 
-// isPrivateIP checks if an IP address is in a private range
-func isPrivateIP(ip net.IP) bool {
-	if ip == nil {
-		return false
-	}
-
-	// Check for loopback
-	if ip.IsLoopback() {
-		return true
-	}
-
-	// Check for link-local
-	if ip.IsLinkLocalUnicast() || ip.IsLinkLocalMulticast() {
-		return true
-	}
-
-	// Check for private ranges (RFC 1918)
-	privateBlocks := []string{
-		"10.0.0.0/8",
-		"172.16.0.0/12",
-		"192.168.0.0/16",
-		"169.254.0.0/16", // AWS metadata endpoint range
-		"127.0.0.0/8",    // Loopback
-		"::1/128",        // IPv6 loopback
-		"fc00::/7",       // IPv6 unique local
-		"fe80::/10",      // IPv6 link local
-	}
-
-	for _, block := range privateBlocks {
-		_, cidr, err := net.ParseCIDR(block)
-		if err != nil {
-			continue
-		}
-		if cidr.Contains(ip) {
-			return true
-		}
-	}
-
-	return false
-}
-
-// validateWebhookHeaders validates that custom webhook headers are safe
-// This prevents HTTP header injection attacks
-func validateWebhookHeaders(headers map[string]string) error {
-	// Blocklist of headers that shouldn't be overridden
-	blockedHeaders := map[string]bool{
-		"content-length":      true,
-		"host":                true,
-		"transfer-encoding":   true,
-		"connection":          true,
-		"keep-alive":          true,
-		"proxy-authenticate":  true,
-		"proxy-authorization": true,
-		"te":                  true,
-		"trailers":            true,
-		"upgrade":             true,
-	}
-
-	for key, value := range headers {
-		lowerKey := strings.ToLower(key)
-
-		// Check for blocked headers
-		if blockedHeaders[lowerKey] {
-			return fmt.Errorf("header '%s' is not allowed to be overridden", key)
-		}
-
-		// Check for CRLF injection in header name
-		if strings.ContainsAny(key, "\r\n") {
-			return fmt.Errorf("header name '%s' contains invalid characters", key)
-		}
-
-		// Check for CRLF injection in header value
-		if strings.ContainsAny(value, "\r\n") {
-			return fmt.Errorf("header value for '%s' contains invalid characters", key)
-		}
-
-		// Limit header value length
-		if len(value) > 8192 {
-			return fmt.Errorf("header value for '%s' exceeds maximum length of 8192 bytes", key)
-		}
-	}
-
-	return nil
-}
-
-// validateWebhookURL validates that a webhook URL is safe to call
-// This prevents SSRF attacks by blocking internal/private IP addresses
-func validateWebhookURL(webhookURL string) error {
-	// Parse the URL
-	parsedURL, err := url.Parse(webhookURL)
-	if err != nil {
-		return fmt.Errorf("invalid URL: %w", err)
-	}
-
-	// Only allow HTTP and HTTPS schemes
-	if parsedURL.Scheme != "http" && parsedURL.Scheme != "https" {
-		return fmt.Errorf("URL scheme must be http or https, got: %s", parsedURL.Scheme)
-	}
-
-	// Get hostname
-	hostname := parsedURL.Hostname()
-	if hostname == "" {
-		return fmt.Errorf("URL must have a hostname")
-	}
-
-	// Check for localhost variants
-	lowerHost := strings.ToLower(hostname)
-	if lowerHost == "localhost" || lowerHost == "ip6-localhost" {
-		return fmt.Errorf("localhost URLs are not allowed")
-	}
-
-	// Check for common internal hostnames
-	blockedHostnames := []string{
-		"metadata.google.internal",
-		"metadata",
-		"instance-data",
-		"kubernetes.default",
-		"kubernetes.default.svc",
-	}
-	for _, blocked := range blockedHostnames {
-		if lowerHost == blocked || strings.HasSuffix(lowerHost, "."+blocked) {
-			return fmt.Errorf("internal hostname '%s' is not allowed", hostname)
-		}
-	}
-
-	// Resolve the hostname and check if it resolves to a private IP
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-	resolver := net.Resolver{}
-	ips, err := resolver.LookupIPAddr(ctx, hostname)
-	if err != nil {
-		// If DNS lookup fails, we can't verify - block it to be safe
-		return fmt.Errorf("failed to resolve hostname: %w", err)
-	}
-
-	for _, ip := range ips {
-		if isPrivateIP(ip.IP) {
-			return fmt.Errorf("URL resolves to private IP address %s which is not allowed", ip.IP.String())
-		}
-	}
-
-	return nil
-}
-
 // NewWebhookService creates a new webhook service
 func NewWebhookService(db *database.Connection) *WebhookService {
 	return &WebhookService{
@@ -248,68 +95,6 @@ func NewWebhookService(db *database.Connection) *WebhookService {
 			Timeout: 30 * time.Second,
 		},
 	}
-}
-
-// parseTableReference splits a table reference into schema and table name
-// e.g.:
-//   - "auth.users" -> ("auth", "users")
-//   - "ai.documents" -> ("ai", "documents")
-//   - "users" -> ("auth", "users") - defaults to auth schema for backward compatibility
-//
-// For AI schema tables, always use the full reference "ai.documents", "ai.chunks", etc.
-func parseTableReference(tableRef string) (schema, table string) {
-	if idx := strings.Index(tableRef, "."); idx > 0 {
-		return tableRef[:idx], tableRef[idx+1:]
-	}
-	// Default to auth schema for backward compatibility
-	// For AI schema tables, use explicit "ai.documents" format
-	return "auth", tableRef
-}
-
-// ManageTriggersForWebhook ensures database triggers exist for all tables monitored by this webhook
-func (s *WebhookService) ManageTriggersForWebhook(ctx context.Context, events []EventConfig) error {
-	for _, event := range events {
-		if event.Table == "*" {
-			continue // Wildcard doesn't need specific trigger
-		}
-		schema, table := parseTableReference(event.Table)
-		if err := s.incrementTableCount(ctx, schema, table); err != nil {
-			return fmt.Errorf("failed to create trigger for %s.%s: %w", schema, table, err)
-		}
-	}
-	return nil
-}
-
-// CleanupTriggersForWebhook decrements reference counts for monitored tables
-func (s *WebhookService) CleanupTriggersForWebhook(ctx context.Context, events []EventConfig) error {
-	for _, event := range events {
-		if event.Table == "*" {
-			continue
-		}
-		schema, table := parseTableReference(event.Table)
-		if err := s.decrementTableCount(ctx, schema, table); err != nil {
-			log.Error().Err(err).Str("schema", schema).Str("table", table).Msg("Failed to decrement table count")
-		}
-	}
-	return nil
-}
-
-// incrementTableCount calls the database function to increment webhook count for a table
-func (s *WebhookService) incrementTableCount(ctx context.Context, schema, table string) error {
-	query := `SELECT auth.increment_webhook_table_count($1, $2)`
-	return database.WrapWithServiceRole(ctx, s.db, func(tx pgx.Tx) error {
-		_, err := tx.Exec(ctx, query, schema, table)
-		return err
-	})
-}
-
-// decrementTableCount calls the database function to decrement webhook count for a table
-func (s *WebhookService) decrementTableCount(ctx context.Context, schema, table string) error {
-	query := `SELECT auth.decrement_webhook_table_count($1, $2)`
-	return database.WrapWithServiceRole(ctx, s.db, func(tx pgx.Tx) error {
-		_, err := tx.Exec(ctx, query, schema, table)
-		return err
-	})
 }
 
 // Create creates a new webhook
@@ -720,298 +505,6 @@ func (s *WebhookService) Delete(ctx context.Context, id uuid.UUID) error {
 	}
 
 	return nil
-}
-
-// Deliver sends a webhook payload to the configured URL
-func (s *WebhookService) Deliver(ctx context.Context, webhook *Webhook, payload *WebhookPayload) error {
-	payloadJSON, err := json.Marshal(payload)
-	if err != nil {
-		return fmt.Errorf("failed to marshal payload: %w", err)
-	}
-
-	// Send HTTP request synchronously and return error if it fails
-	// The trigger service will handle retries via webhook_events table
-	return s.sendWebhookSync(ctx, webhook, payloadJSON)
-}
-
-// sendWebhookSync sends an HTTP request synchronously and returns any error
-func (s *WebhookService) sendWebhookSync(ctx context.Context, webhook *Webhook, payloadJSON []byte) error {
-	// SECURITY FIX: Validate webhook URL at request time to prevent DNS rebinding attacks
-	// An attacker could create a webhook with a URL that initially resolves to a public IP,
-	// then change the DNS to point to a private IP (e.g., 169.254.169.254 for cloud metadata)
-	if !s.AllowPrivateIPs() {
-		if err := validateWebhookURL(webhook.URL); err != nil {
-			return fmt.Errorf("webhook URL validation failed (possible DNS rebinding attack): %w", err)
-		}
-	}
-
-	req, err := http.NewRequestWithContext(ctx, "POST", webhook.URL, bytes.NewReader(payloadJSON))
-	if err != nil {
-		return fmt.Errorf("failed to create request: %w", err)
-	}
-
-	// Set headers
-	req.Header.Set("Content-Type", "application/json")
-	req.Header.Set("User-Agent", "Fluxbase-Webhooks/1.0")
-
-	// Add custom headers
-	for key, value := range webhook.Headers {
-		req.Header.Set(key, value)
-	}
-
-	// Add HMAC signature with timestamp if secret is provided
-	// Format: t=timestamp,v1=signature
-	// This enables replay protection and is similar to Stripe's webhook signing
-	if webhook.Secret != nil && *webhook.Secret != "" {
-		timestamp := time.Now().Unix()
-		signature := generateTimestampedSignature(payloadJSON, *webhook.Secret, timestamp)
-		signatureHeader := fmt.Sprintf("t=%d,v1=%s", timestamp, signature)
-		req.Header.Set("X-Fluxbase-Signature", signatureHeader)
-		// Also keep legacy header for backwards compatibility
-		legacySignature := s.generateSignature(payloadJSON, *webhook.Secret)
-		req.Header.Set("X-Webhook-Signature", legacySignature)
-	}
-
-	// Send request with timeout
-	client := &http.Client{
-		Timeout: time.Duration(webhook.TimeoutSeconds) * time.Second,
-	}
-
-	resp, err := client.Do(req)
-	if err != nil {
-		return fmt.Errorf("failed to send request: %w", err)
-	}
-	defer func() { _ = resp.Body.Close() }()
-
-	// Check status code
-	if resp.StatusCode < 200 || resp.StatusCode >= 300 {
-		body, _ := io.ReadAll(resp.Body)
-		return fmt.Errorf("HTTP %d: %s", resp.StatusCode, string(body))
-	}
-
-	return nil
-}
-
-// sendWebhook sends the actual HTTP request (runs asynchronously).
-// Note: Currently unused but kept for potential future async webhook delivery implementation.
-/*
-func (s *WebhookService) sendWebhook(ctx context.Context, deliveryID uuid.UUID, webhook *Webhook, payloadJSON []byte) {
-	// SECURITY FIX: Validate webhook URL at request time to prevent DNS rebinding attacks
-	if !s.AllowPrivateIPs() {
-		if err := validateWebhookURL(webhook.URL); err != nil {
-			s.markDeliveryFailed(ctx, deliveryID, 0, nil, fmt.Sprintf("webhook URL validation failed (possible DNS rebinding): %v", err))
-			return
-		}
-	}
-
-	// Create HTTP request
-	req, err := http.NewRequestWithContext(ctx, "POST", webhook.URL, bytes.NewReader(payloadJSON))
-	if err != nil {
-		s.markDeliveryFailed(ctx, deliveryID, 0, nil, fmt.Sprintf("failed to create request: %v", err))
-		return
-	}
-
-	// Set headers
-	req.Header.Set("Content-Type", "application/json")
-	req.Header.Set("User-Agent", "Fluxbase-Webhooks/1.0")
-
-	// Add custom headers
-	for key, value := range webhook.Headers {
-		req.Header.Set(key, value)
-	}
-
-	// Add HMAC signature with timestamp if secret is provided
-	if webhook.Secret != nil && *webhook.Secret != "" {
-		timestamp := time.Now().Unix()
-		signature := generateTimestampedSignature(payloadJSON, *webhook.Secret, timestamp)
-		signatureHeader := fmt.Sprintf("t=%d,v1=%s", timestamp, signature)
-		req.Header.Set("X-Fluxbase-Signature", signatureHeader)
-		// Also keep legacy header for backwards compatibility
-		legacySignature := s.generateSignature(payloadJSON, *webhook.Secret)
-		req.Header.Set("X-Webhook-Signature", legacySignature)
-	}
-
-	// Send request with timeout
-	client := &http.Client{
-		Timeout: time.Duration(webhook.TimeoutSeconds) * time.Second,
-	}
-
-	resp, err := client.Do(req)
-	if err != nil {
-		s.markDeliveryFailed(ctx, deliveryID, 0, nil, fmt.Sprintf("failed to send request: %v", err))
-		return
-	}
-	defer func() { _ = resp.Body.Close() }()
-
-	// Read response body
-	body, _ := io.ReadAll(resp.Body)
-	bodyStr := string(body)
-
-	// Check status code
-	if resp.StatusCode >= 200 && resp.StatusCode < 300 {
-		s.markDeliverySuccess(ctx, deliveryID, resp.StatusCode, &bodyStr)
-	} else {
-		s.markDeliveryFailed(ctx, deliveryID, resp.StatusCode, &bodyStr, fmt.Sprintf("HTTP %d", resp.StatusCode))
-	}
-}
-*/
-
-// generateSignature generates HMAC SHA256 signature (legacy, without timestamp)
-func (s *WebhookService) generateSignature(payload []byte, secret string) string {
-	mac := hmac.New(sha256.New, []byte(secret))
-	mac.Write(payload)
-	return hex.EncodeToString(mac.Sum(nil))
-}
-
-// generateTimestampedSignature generates HMAC SHA256 signature with timestamp
-// The signature is computed over: timestamp + "." + payload
-// This prevents replay attacks by including the timestamp in the signed data
-func generateTimestampedSignature(payload []byte, secret string, timestamp int64) string {
-	// Create the signed payload: timestamp.payload
-	signedPayload := fmt.Sprintf("%d.%s", timestamp, string(payload))
-
-	mac := hmac.New(sha256.New, []byte(secret))
-	mac.Write([]byte(signedPayload))
-	return hex.EncodeToString(mac.Sum(nil))
-}
-
-// WebhookSignature represents a parsed webhook signature header
-type WebhookSignature struct {
-	Timestamp  int64
-	Signatures []string
-}
-
-// ParseWebhookSignature parses an X-Fluxbase-Signature header
-// Format: t=timestamp,v1=signature[,v1=signature2...]
-// Example: t=1234567890,v1=abc123def456
-func ParseWebhookSignature(header string) (*WebhookSignature, error) {
-	if header == "" {
-		return nil, fmt.Errorf("empty signature header")
-	}
-
-	sig := &WebhookSignature{}
-
-	parts := strings.Split(header, ",")
-	for _, part := range parts {
-		kv := strings.SplitN(strings.TrimSpace(part), "=", 2)
-		if len(kv) != 2 {
-			continue
-		}
-
-		key := strings.TrimSpace(kv[0])
-		value := strings.TrimSpace(kv[1])
-
-		switch key {
-		case "t":
-			ts, err := strconv.ParseInt(value, 10, 64)
-			if err != nil {
-				return nil, fmt.Errorf("invalid timestamp: %w", err)
-			}
-			sig.Timestamp = ts
-		case "v1":
-			sig.Signatures = append(sig.Signatures, value)
-		}
-	}
-
-	if sig.Timestamp == 0 {
-		return nil, fmt.Errorf("missing timestamp in signature")
-	}
-	if len(sig.Signatures) == 0 {
-		return nil, fmt.Errorf("missing signature value")
-	}
-
-	return sig, nil
-}
-
-// VerifyWebhookSignature verifies a webhook signature
-// Parameters:
-//   - payload: the raw request body
-//   - header: the X-Fluxbase-Signature header value
-//   - secret: the webhook secret
-//   - tolerance: maximum age of the signature (recommended: 5 minutes)
-//
-// Returns nil if signature is valid, error otherwise
-func VerifyWebhookSignature(payload []byte, header, secret string, tolerance time.Duration) error {
-	sig, err := ParseWebhookSignature(header)
-	if err != nil {
-		return fmt.Errorf("failed to parse signature: %w", err)
-	}
-
-	// Check timestamp is not too old (replay protection)
-	signedAt := time.Unix(sig.Timestamp, 0)
-	if time.Since(signedAt) > tolerance {
-		return fmt.Errorf("signature timestamp too old (signed at %v, tolerance %v)", signedAt, tolerance)
-	}
-
-	// Check timestamp is not in the future (clock skew protection)
-	if signedAt.After(time.Now().Add(tolerance)) {
-		return fmt.Errorf("signature timestamp in the future")
-	}
-
-	// Compute expected signature
-	expectedSig := generateTimestampedSignature(payload, secret, sig.Timestamp)
-
-	// Compare signatures (constant time comparison to prevent timing attacks)
-	for _, providedSig := range sig.Signatures {
-		if hmac.Equal([]byte(expectedSig), []byte(providedSig)) {
-			return nil
-		}
-	}
-
-	return fmt.Errorf("signature mismatch")
-}
-
-// CreateDeliveryRecord creates a delivery record before attempting delivery
-func (s *WebhookService) CreateDeliveryRecord(ctx context.Context, webhookID uuid.UUID, event string, payload []byte, attempt int, tenantID string) (uuid.UUID, error) {
-	query := `
-		INSERT INTO auth.webhook_deliveries (webhook_id, event, payload, status, attempt)
-		VALUES ($1, $2, $3, 'pending', $4)
-		RETURNING id
-	`
-
-	var deliveryID uuid.UUID
-	err := database.WrapWithServiceRoleAndTenant(ctx, s.db, tenantID, func(tx pgx.Tx) error {
-		return tx.QueryRow(ctx, query, webhookID, event, payload, attempt).Scan(&deliveryID)
-	})
-	if err != nil {
-		return uuid.Nil, fmt.Errorf("failed to create delivery record: %w", err)
-	}
-
-	return deliveryID, nil
-}
-
-// markDeliverySuccess marks a delivery as successful
-func (s *WebhookService) markDeliverySuccess(ctx context.Context, deliveryID uuid.UUID, statusCode int, responseBody *string) {
-	query := `
-		UPDATE auth.webhook_deliveries
-		SET status = 'success', status_code = $1, response_body = $2, delivered_at = NOW()
-		WHERE id = $3
-	`
-
-	err := database.WrapWithServiceRole(ctx, s.db, func(tx pgx.Tx) error {
-		_, err := tx.Exec(ctx, query, statusCode, responseBody, deliveryID)
-		return err
-	})
-	if err != nil {
-		log.Error().Err(err).Str("delivery_id", deliveryID.String()).Msg("Failed to mark delivery as success")
-	}
-}
-
-// markDeliveryFailed marks a delivery as failed
-func (s *WebhookService) markDeliveryFailed(ctx context.Context, deliveryID uuid.UUID, statusCode int, responseBody *string, errorMsg string) {
-	query := `
-		UPDATE auth.webhook_deliveries
-		SET status = 'failed', status_code = $1, response_body = $2, error = $3
-		WHERE id = $4
-	`
-
-	err := database.WrapWithServiceRole(ctx, s.db, func(tx pgx.Tx) error {
-		_, err := tx.Exec(ctx, query, statusCode, responseBody, errorMsg, deliveryID)
-		return err
-	})
-	if err != nil {
-		log.Error().Err(err).Str("delivery_id", deliveryID.String()).Msg("Failed to mark delivery as failed")
-	}
 }
 
 // ListDeliveries lists webhook deliveries
